src/requirements.txt
absl-py==0.6.1
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
django-nose
boto==2.49.0
boto3==1.9.37
botocore==1.12.37
bz2file==0.98
certifi==2018.10.15
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==4.5.1
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.8.2
Django==2.0.8
django-cors-headers==2.4.0
django-jet==1.0.8
djangorestframework==3.9.0
docopt==0.6.2
docutils==0.14
Faker==0.9.1
gast==0.2.0
gensim==3.6.0
grpcio==1.16.0
h5py==2.8.0
hdfs==2.1.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
MarkupSafe==1.1.0
mixer==6.1.3
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nltk==3.3
numpy==1.15.4
pandas==0.23.4
plac==0.9.6
pluggy==0.8.0
preshed==2.0.1
protobuf==3.6.1
py==1.7.0
py4j==0.10.7
Pyphen==0.9.5
pytest==3.10.0
pytest-cov==2.6.0
pytest-django==3.4.3
python-dateutil==2.7.5
pytz==2018.7
pywebhdfs==0.4.1
PyYAML==3.13
regex==2018.1.10
repoze.lru==0.7
requests==2.20.0
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
six==1.12.0
sklearn==0.0
smart-open==1.7.1
spacy==2.0.16
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
text-unidecode==1.2
textstat==0.5.1
thinc==6.12.0
toolz==0.9.0
tqdm==4.28.1
ujson==1.35
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
xgboost==0.80
newspaper3k
newsapi-python